Fixed issues found during testing of BulkCreateMonumentSuggestion API flow
